+++ Sources/Recap/Public/Feature.swift
@@ -7,18 +7,24 @@ public struct Feature: Identifiable, Equatable {
     public let title: String
     public let description: String
     public let symbolName: String
-    public let hexColor: String
+    public let colorRepresentation: String
 
-    public init(title: String, description: String, symbolName: String, hexColor: String) {
+    public init(title: String, description: String, symbolName: String, colorRepresentation: String) {
         self.title = title
         self.description = description
         self.symbolName = symbolName
-        self.hexColor = hexColor
+        self.colorRepresentation = colorRepresentation
     }
 }
 
 internal extension Feature {
     var color: Color {
-        Color(hex: self.hexColor)
+        if self.colorRepresentation.hasPrefix("#") {
+            return Color(hex: self.colorRepresentation)
+        } else if let color = Color(systemName: self.colorRepresentation) {
+            return color
+        } else {
+            return .black
+        }
     }
 }